module: Name the module after the tool's name
The cmd tool is compiled into a binary called 'uhhm' by default.
However, the module was called 'uhhmm' just to give the name an extra
bit of thoughtfulness. That didn't sound quite much like a good practice
though.

So I'm calling the module uhhm just to reduce the entropy a bit.

Signed-off-by: Beñat Gartzia Arruabarrena <begartzia@gmail.com>